done git add .

+++ 0x00-hello_world/6-size.c
@@ -11,11 +11,11 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of a char: %Iu byte(s)\n", sizeof(a));
-	printf("Size of an int: %Iu byte(s)\n", sizeof(b));
-	printf("Size of a long int: %Iu byte(s)\n", sizeof(c));
-	printf("Size Of a long long int: %Iu byte(s)\n", sizeof(d));
-	printf("Size of a float: %Iu byte(s)\n", sizeof(e));
+	printf("Size of a char: %Iu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %Iu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a long int: %Iu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size Of a long long int: %Iu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a float: %Iu byte(s)\n", (unsigned long)sizeof(e));
 
 	return (0);
 	}
